+++ RELEASES.md
@@ -1,5 +1,23 @@
 # Enso Next
 
+# Enso 0.2.17 (2021-07-28)
+
+## Interpreter/Runtime
+
+- Added support for documenting modules directly
+  ([#1900](https://github.com/enso-org/enso/pull/1900)).
+
+## Tooling
+
+- Added support for creating projects from a template
+  ([#1902](https://github.com/enso-org/enso/pull/1902)).
+
+## Known Issues
+
+- This is a beta release, so please see the
+  [issue tracker](https://github.com/enso-org/enso/issues?q=is%3Aissue+is%3Aopen+created%3A%3C2021-07-28)
+  for issues opened before the release date.
+
 # Enso 0.2.16 (2021-07-23)
 
 ## Interpreter/Runtime